First cut at standard way of making dynamically loadable extensions on UNIX. Examples on how to test:
 ./configure --with-xml              static
 ./configure --with-xml=/opt         static
 ./configure --with-xml=shared       shared
 ./configure --with-xml=shared,/opt  shared

The difference between these two is that when the extension is shared, it
is not merged into libphpext.a.  The shared extension is currently always
built.  I can't find a way to do just one or the other with automake/libtool,
if someone has a clever idea, please come forward. :-)

"make install" installs the .so (as well as a lot of other cruft) in
$prefix/lib/php.


git-svn-id: http://svn.php.net/repository/php/php-src/trunk@11118 c90b9560-bf6c-de11-be94-00142212c4b1