@- Fixed signal() bug that could cause the Apache master process to @  die. (Thies)
the problem was that we *did* register the signal in the apache
master process as well.
php_module_startup() -> REGISTER_INI_ENTRIES() -> OnChangeMaxExecutionTime() -> php_set_timeout().

from now on the .ini mechanism only sets PG(max_execution_time) and that gets
"converted" to a timeout-signal() in php_request_startup()!

we still need to think about the ZTS timeout mechanism.


git-svn-id: http://svn.php.net/repository/php/php-src/trunk@19260 c90b9560-bf6c-de11-be94-00142212c4b1