Improve the general behaviour of stream_gets and fix its semantics with regard to sockets. The behaviour should be aligned with PHP 4.2 now. This has been verified to some degree.
If the underlying stream operations block when no new data is readable,
we need to take extra precautions.

If there is buffered data available, we check for a EOL. If it exists,
we pass the data immediately back to the caller. This saves a call
to the read implementation and will not block where blocking
is not necessary at all.

If the stream buffer contains more data than the caller requested,
we can also avoid that costly step and simply return that data.


git-svn-id: http://svn.php.net/repository/php/php-src/trunk@98166 c90b9560-bf6c-de11-be94-00142212c4b1