Add --enable-snapshot-build configure option for win32 build. Using it will attempt to switch on any configure options that were left to their default value of "no".
Add a "snap" target to the makefile which will run a sub-process
to make each SAPI and EXT independently, so that the whole process
does not bail out if a particular extension is broken.

Add a way to specify dependencies between extensions in the
config.w32 files so that the makefile will include appropriate rules
and linkage.


git-svn-id: http://svn.php.net/repository/php/php-src/trunk@147160 c90b9560-bf6c-de11-be94-00142212c4b1