Tidy up "nmake clean" for people that don't have a PECL checkout.
Be paranoid when building a snapshot: if the module is not a core
module (eg: it comes from outside of ext or sapi) and it defaults to "yes",
then force it to become shared.  This will prevent a pecl ext from
accidentally being compiled statically into the core, and prevent that
ext from breaking the core build.  You can still manually force a static
build by explicitly specifying the args for that extension on your configure
line.


git-svn-id: http://svn.php.net/repository/php/php-src/trunk@151233 c90b9560-bf6c-de11-be94-00142212c4b1