Take a blind stab at implementing scrollable cursors for pgsql. We allocate a unique cursor name for each statement, so that we don't interfere with other open statement handles on the same dbh.
Note, however, that we force a new transaction for each open scrollable cursor
(postgres requires cursors to be used inside a transaction).  This is okay,
except for the case where a scrollable cursor is opened, an update is made and
the cursor is closed; closing the cursor commits the transaction that was begun
when it was opened.

It might well be better to avoid the transaction in PDO and force the user to
be aware of the requirements of cursors and explicitly initiate the transaction
themselves.

This is all untested code; it compiles and looks like it will work, but I
encourage someone with a real postgres setup to actually sit down and try to
use it.


git-svn-id: http://svn.php.net/repository/php/php-src/trunk@177111 c90b9560-bf6c-de11-be94-00142212c4b1