add a bit of a hack to cater for pgsql prepared statements.
These are effectively named statements with strong constraints on the naming
format.  We cater for this in a fairly generic way: allow a driver to replace
the format string we use to generate names from positional parameters.  In
addition, if that format is set, we always force a rewrite from regular names
to the strongly enforced names.


git-svn-id: http://svn.php.net/repository/php/php-src/trunk@190258 c90b9560-bf6c-de11-be94-00142212c4b1