[DOC] Added mail logging functionality that allows logging of mail sent via mail() function
THe logging functionality is disabled by default but can be enabled on a
per-directory or globaly via 2 INI settings.

The mail.log directive allows you to specify the file where each call to
mail() will be logged. The log file will contain the path and line # of the
calling script in addition to all of the headers indicated by the user.

The mail.add_x_header directive will introduce a X-PHP-Originating-Script
header that will contain the file name (no path) of the calling script and
the uid of the script. This combination should be sufficient for the admin
to locate the sending script.


git-svn-id: http://svn.php.net/repository/php/php-src/branches/PHP_5_3@273168 c90b9560-bf6c-de11-be94-00142212c4b1